fix laravel command not working
You need to export $PATH after the script is finished...but it is individual and you need to check it like this: First make sure that composer is working where ever you type composer (in any directory). If it works then just export laravel => first check where is it; It must be in ~/.config/composer/vendor/bin or in ~/.composer/vendor/bin. Export it with this command accordingly where the "laravel" actually was: export $PATH=~/.config/composer/vendor/bin:$PATH